+++ addressbook-web-tests/src/test/java/ru/ls/qa/school/addressbook/appmanager/ApplicationManager.java
@@ -1,38 +1,24 @@
 package ru.ls.qa.school.addressbook.appmanager;
 
 import com.codeborne.selenide.Configuration;
-import ru.ls.qa.school.addressbook.model.GroupData;
 
 import java.util.Map;
 
-import static com.codeborne.selenide.Selectors.byLinkText;
-import static com.codeborne.selenide.Selectors.byName;
 import static com.codeborne.selenide.Selenide.*;
 
 public class ApplicationManager {
     private final GroupHelper groupHelper = new GroupHelper();
     private final NavigationHelper navigationHelper = new NavigationHelper();
     private final ContactHelper contactHelper = new ContactHelper();
+    private final SessionHelper sessionHelper = new SessionHelper();
     public Map<String, Object> vars;
 
-
-    public  void logout() {
-        $(byLinkText("Logout")).click();
-    }
-
-    public static void login(String username, String password) {
-        open("http://localhost/addressbook");
-        $(byName("user")).click();
-        $(byName("user")).val(username);
-        $(byName("pass")).val(password);
-        $("input:nth-child(7)").click();
-    }
 
     public void init() {
         Configuration.browser = "chrome";
         Configuration.holdBrowserOpen = true;
         open("http://localhost/addressbook");
-        ApplicationManager.login("admin", "secret");
+        SessionHelper.login("admin", "secret");
     }
 
     public void stop() {
@@ -49,5 +35,9 @@ public class ApplicationManager {
 
     public ContactHelper getContactHelper() {
         return contactHelper;
+    }
+
+    public SessionHelper getSessionHelper() {
+        return sessionHelper;
     }
 }
