+++ addressbook-web-tests/src/test/java/ru/ls/qa/school/addressbook/tests/NewContactTest.java
@@ -1,5 +1,4 @@
-package ru.ls.qa.school.addressbook.tests;// Generated by Selenium IDE
-
+package ru.ls.qa.school.addressbook.tests;
 import org.junit.jupiter.api.Test;
 import ru.ls.qa.school.addressbook.model.ContactData;
 
@@ -15,7 +14,7 @@ public class NewContactTest extends TestBase {
                 "Moscow"));
         app.getContactHelper().submitCreationNewContact();
         app.getNavigationHelper().returnToMainPage();
-        app.logout();
+        app.getSessionHelper().logout();
     }
 }
